backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
langchain>=0.3
langchain-openai>=0.3
langchain-postgres>=0.0.12
langchain-community>=0.3
langchain-text-splitters>=0.3
redis>=5.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9
sse-starlette>=2.0
python-docx>=1.1
docx2txt>=0.8
pymupdf>=1.25
ragas>=0.2
langsmith>=0.2
structlog>=24.0
